backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas>=2.2.0
openpyxl==3.1.2
deep-translator
google-generativeai
python-dotenv
websockets>=12.0
openai-whisper
numpy
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

